requirements.txt
# Core ML and Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
tensorflow>=2.8.0
torch>=1.11.0
transformers>=4.15.0

# Web Framework and API
fastapi>=0.75.0
uvicorn>=0.17.0
pydantic>=1.9.0
python-multipart>=0.0.5

# Email Processing
imaplib2>=3.6
email-validator>=1.1.3
beautifulsoup4>=4.10.0
lxml>=4.7.0

# Feature Extraction and Web Scraping
requests>=2.27.0
urllib3>=1.26.0
python-whois>=0.7.3
tldextract>=3.2.0

# Database and Caching
sqlalchemy>=1.4.0
alembic>=1.7.0
redis>=4.1.0
psycopg2-binary>=2.9.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5

# Monitoring and Logging
prometheus-client>=0.13.0
structlog>=21.5.0
sentry-sdk>=1.5.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
httpx>=0.22.0

# Development Tools
black>=22.0.0
flake8>=4.0.0
mypy>=0.931
pre-commit>=2.17.0

# Security Tools
bandit>=1.7.4
safety>=2.3.0
pip-audit>=2.4.0
detect-secrets>=1.4.0

# Deployment
docker>=5.0.0
kubernetes>=21.7.0
